+++ changes.md
@@ -6,6 +6,7 @@
   item being dragged. Is `"dragged"` by default.
 - New property for sortable pattern, `drop`, a Javascript callback function to
   be called when an item is dropped.
+- Form with pat-autosubmit doesn't get notified when injected inputs change.
 
 ## 2.0.12 - Oct. 9, 2015